[WK2] [GTK] LayerTreeHostGtk: don't set m_isValid to false if glContext() returns null https://bugs.webkit.org/show_bug.cgi?id=120892
Reviewed by Martin Robinson.

If the GL context cannot be created then m_isValid is set to false
during the initialization of LayerTreeHostGtk.

This is not really necessary since the rest of the code already
deals with that situation, so it doesn't have any effect other
than breaking the assertion in invalidate().

* WebProcess/WebPage/gtk/LayerTreeHostGtk.cpp:
(WebKit::LayerTreeHostGtk::initialize):

Identifier: 139091@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@155533 268f45cc-cd09-0410-ab3c-d52691b4dbfc