REGRESSION (143483): overflow:hidden doesn't quash big repaints from text-indent: -9999px https://bugs.webkit.org/show_bug.cgi?id=121137
Source/WebCore:

Reviewed by Ryosuke Niwa.

The change in r143483 made it so that negative text-indent gets added
to visual oveflow even when we have overflow:hidden, which is really bad
because it causes large repaints (and large layers) for a common-on-the-web
pattern.

Revert the change, and add a new test to make sure it doesn't regress.

I can't reproduce the bug it was trying to fix with the change reverted.

Tests: compositing/geometry/negative-text-indent-with-overflow-hidden-layer.html
       fast/repaint/negative-text-indent-with-overflow-hidden.html

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::computeOverflow):

LayoutTests:

Reviewed by Ryosuke Niwa.

Text-based repaint test, and test for compositing layer geometry.

* compositing/geometry/negative-text-indent-with-overflow-hidden-layer-expected.txt: Added.
* compositing/geometry/negative-text-indent-with-overflow-hidden-layer.html: Added.
* fast/repaint/negative-text-indent-with-overflow-hidden-expected.txt: Added.
* fast/repaint/negative-text-indent-with-overflow-hidden.html: Added.

Identifier: 139104@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@155546 268f45cc-cd09-0410-ab3c-d52691b4dbfc