[CSSRegions] Failed to retrieve named flow ranges for content with inline svg https://bugs.webkit.org/show_bug.cgi?id=121489
Reviewed by David Hyatt.

Source/WebCore:

The inline SVG element behaves like a replaced element in HTML context. When computing the ranges
that are displayed inside a region for a named flow that contains an inline SVG element, we have
to treat the SVG as a replaced and run the checks only for the SVG root element, skipping the
other SVG elements inside.

Test: fast/regions/get-region-flow-ranges-inline-svg.html

* rendering/RenderNamedFlowThread.cpp:
(WebCore::nextNodeInsideContentNode): Add function to retrieve the next node in traversal
under the contentNode subtree that should be used for getRanges inspection.
(WebCore::RenderNamedFlowThread::getRanges): Replace NodeTravesal::next with nextNodeInsideContentNode.

LayoutTests:

* fast/regions/get-region-flow-ranges-inline-svg-expected.txt: Added.
* fast/regions/get-region-flow-ranges-inline-svg.html: Added.


Identifier: 139597@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156082 268f45cc-cd09-0410-ab3c-d52691b4dbfc