Optimize fetching the Node for never-anonymous renderers. <https://webkit.org/b/121627>
Reviewed by Antti Koivisto.

Add a RenderObject::nodeForNonAnonymous() that can be used instead
of node() when a renderer is guaranteed non-anonymous.

Use this to implement reference element getters for the various
renderers that are known at compile-time to correspond to a DOM node.

Also renamed existingElement() => element() since using an alternate
name for a reference overload isn't actually necessary.


Identifier: 139670@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156155 268f45cc-cd09-0410-ab3c-d52691b4dbfc