REGRESSION (r156140): Srcset tests are frequently crashing https://bugs.webkit.org/show_bug.cgi?id=121695
Reviewed by Dean Jackson.

Returning a string created without copying bytes is not safe. It used to be OK
because a new string was immediately created by decodeURLEscapeSequences().
But even that was not great, because decodeURLEscapeSequences() could potentially
return the same string, not a deep copy, if we decided to optimize it like that.

Also made a number of drive-by style fixes.
- It's URL, not Url.
- It's srcset, not srcSet.
-  We don't add ".0" in floating point value initializers. It's particularly misleading
to initialize a float with 1.0, which is a double value.
- Renamed srcSetLength to srcsetAttributeLength to match srcsetAttribute variable
whose length it caches.

* html/parser/HTMLParserIdioms.cpp:
(WebCore::parseImagesWithScaleFromSrcsetAttribute):
(WebCore::bestFitSourceForImageAttributes):



Identifier: 139698@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156183 268f45cc-cd09-0410-ab3c-d52691b4dbfc