Make extract-localizable-strings available to WebCore clients https://bugs.webkit.org/show_bug.cgi?id=121699
Reviewed by Darin Adler.

Source/WebCore:

* WebCore.xcodeproj/project.pbxproj: Added a Scripts group with a reference to
extract-localizable-strings.pl. Added a Copy Scripts build phase to the WebCore target that
copies the script into PrivateHeaders/Scripts in the framework.
* extract-localizable-strings.pl: Copied from Tools/Scripts/extract-localizable-strings.

Tools:

* Scripts/extract-localizable-strings: Copied to
Soure/WebCore/extract-localizaebl-strings.pl, and changed this into a shell script that
executes the script from its new location. This is a temporary measure until certain code
that is external to WebKit is adapted to the new location.
* Scripts/update-webkit-localizable-strings: Changed to call extract-localizable-strings.pl
from its new location.


Identifier: 139706@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156191 268f45cc-cd09-0410-ab3c-d52691b4dbfc