Remove RenderObjectChildList https://bugs.webkit.org/show_bug.cgi?id=121771
Reviewed by Darin Adler.

Most RenderElement subclasses have m_children member (>90% of instances on typical page). It is not worth optimising
for memory for the few cases that don't (RenderImage, RenderLineBreak mostly) in expense of performance.

This patch moves child handling to RenderElement. This gets rid of virtual children() calls and simplifies the code.



Identifier: 139765@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156254 268f45cc-cd09-0410-ab3c-d52691b4dbfc