REGRESSION(r154219): 30% more malloc memory used on html5-full-render. <https://webkit.org/b/122034>
Reviewed by Antti Koivisto.

canExecuteScripts() used to rely on Frame::settings() returning null
when the Frame was disconnected from its Page, and giving Frame its
own handle on the Settings broke this mechanism.

This caused us to hold on for some objects for a bit longer than we
needed to in the run-perf-tests harness, showing up as a massive 30%
regression in malloc memory use.

Fixed it by added an explicit null check of Frame::page().


Identifier: 140079@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156582 268f45cc-cd09-0410-ab3c-d52691b4dbfc