Quirksmode: CSS1: WebKit fails dynamic :first-letter test https://bugs.webkit.org/show_bug.cgi?id=15602
Reviewed by David Hyatt.

Source/WebCore:

CSS first-letter property does not work properly when the first letter is changed
by DOM scripting.
This patch allows to check if the existing first-letter is no longer the
first-letter. In this case, it deletes the old first-letter object and creates
a new one. For the remaining text, the oldRemainingText object is used
again for containing the full text(first letter + remaining text).

Test: fast/css/first-letter-block-change.html

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::updateFirstLetter):

LayoutTests:

Add a test case that ensures that CSS first-letter property should work
properly when the first letter is changed by DOM scripting.

* fast/css/first-letter-block-change.html: Added.
* platform/efl/TestExpectations:
* platform/gtk-wk1/fast/css/first-letter-block-change-expected.png: Added.
* platform/gtk-wk2/fast/css/first-letter-block-change-expected.png: Added.
* platform/gtk/fast/css/first-letter-block-change-expected.txt: Added.
* platform/mac/TestExpectations:


Identifier: 140235@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156742 268f45cc-cd09-0410-ab3c-d52691b4dbfc